refactor: USER_QUERY_OPTIONS 통합 및 쿼리 경로 수정
