+++ apps/client/src/shared/api/domain/onboarding/queries.ts
@@ -10,16 +10,9 @@ import {
   UserInfoOptions,
   UserInfoSubmitRequest,
   UserInfoSubmitResponse,
-  UserProfile,
 } from '@shared/api/types/types';
 
-export const USER_QUERY_OPTIONS = {
-  PROFILE: () => {
-    return queryOptions({
-      queryKey: USER_QUERY_KEY.PROFILE(),
-      queryFn: getUserProfile,
-    });
-  },
+export const INSURANCE_QUERY_OPTIONS = {
   JOBS: () => {
     return queryOptions({
       queryKey: USER_QUERY_KEY.JOBS(),
@@ -55,13 +48,6 @@ export const usePostUserInfo = (onSuccessCallback?: () => void) => {
       }
     },
   });
-};
-
-export const getUserProfile = async (): Promise<UserProfile | null> => {
-  const response = await api
-    .get(END_POINT.USER.GET_USER_INFO)
-    .json<UserProfile>();
-  return response;
 };
 
 export const getUserInfoJobs = async (): Promise<UserInfoJobs | null> => {
