+++ apps/client/src/shared/api/domain/report/queries.ts
@@ -2,10 +2,7 @@ import { queryOptions } from '@tanstack/react-query';
 
 import { END_POINT } from '@shared/api/config/end-point';
 import { api } from '@shared/api/config/instance';
-import {
-  INSURANCE_QUERY_KEY,
-  USER_QUERY_KEY,
-} from '@shared/api/keys/query-key';
+import { INSURANCE_QUERY_KEY } from '@shared/api/keys/query-key';
 import {
   InsuranceIpwonReport,
   InsuranceJanghaeReport,
@@ -14,7 +11,6 @@ import {
   InsuranceSamangReport,
   InsuranceSummary,
   InsuranceSusulReport,
-  UserProfile,
 } from '@shared/api/types/types';
 
 export const INSURANCE_QUERY_OPTIONS = {
@@ -59,15 +55,6 @@ export const INSURANCE_QUERY_OPTIONS = {
     return queryOptions({
       queryKey: INSURANCE_QUERY_KEY.REPORT_SECION(reportId, section),
       queryFn: () => getInsuranceSamangReport(reportId, section),
-    });
-  },
-};
-
-export const USER_QUERY_OPTIONS = {
-  PROFILE: () => {
-    return queryOptions({
-      queryKey: USER_QUERY_KEY.PROFILE(),
-      queryFn: getUserProfile,
     });
   },
 };
@@ -146,12 +133,5 @@ export const getInsuranceSamangReport = async (
       searchParams: { section },
     })
     .json<InsuranceSamangReport>();
-  return response;
-};
-
-export const getUserProfile = async (): Promise<UserProfile | null> => {
-  const response = await api
-    .get(END_POINT.USER.GET_USER_INFO)
-    .json<UserProfile>();
   return response;
 };
