+++ apps/client/src/widgets/mypage/components/preview/preview.tsx
@@ -5,13 +5,12 @@ import { useNavigate } from 'react-router';
 import { Tab } from '@bds/ui';
 
 import EmptyPlaceholder from '@widgets/community/components/empty-placeholder/empty-placeholder';
+import CommentPreview from '@widgets/mypage/components/comment-preview/comment-preview';
+import PostPreview from '@widgets/mypage/components/post-preview/post-preview';
 
-import { USER_QUERY_OPTIONS } from '@shared/api/domain/mypage/queries';
+import { USER_QUERY_OPTIONS } from '@shared/api/domain/queries';
 import { useIntersectionObserver } from '@shared/hooks/use-intersection-observer';
 import { routePath } from '@shared/router/path';
-
-import CommentPreview from '../comment-preview/comment-preview';
-import PostPreview from '../post-preview/post-preview';
 
 import * as styles from './preview.css';
